Move agent and redirect handling into separate files
- Create a class for the default Agent so that the agents won't be
  shared globally as that might be unexpected
- Fix Agent defaults
- Move redirect handling into a separate file to allow better
  test coverage
